requirements.txt
Django==2.2.5
django-bootstrap3-datetimepicker-2==2.7.0
django-ckeditor==5.5.0
django-js-asset==1.1.0
